treewide: fix Python lint issues and drop ruff ignore list
Remove the entire lint ignore list from .ruff.toml and fix all
resulting findings across the Python tooling:

- Convert printf-style % formatting to str.format()/f-strings
  (UP031/UP032)
- Replace raise Exception with RuntimeError (TRY002)
- Use module-level loggers instead of the root logger (LOG015)
- Fix undefined names: Python 2 leftovers (xrange, unichr, basestring,
  file, print >>), missing imports, and real typos (F821)
- Restore the clang import probes in parse_c_decl.py that had been
  reduced to no-op "try: pass" blocks, and define the missing
  CodeRejectCatastrophic/ConfigurationRejected classes in pulse2 ppp
- Declare __all__ in re-export __init__ modules (F401)
- Replace bare/blind excepts with specific exception types,
  contextlib.suppress for cleanup paths, and a few explicit
  noqa: BLE001 for intentional catch-alls (E722/BLE001/S110)
- Fix shebang/executable-bit mismatches (EXE001/EXE002)
- Annotate mutable class attributes with ClassVar (RUF012)
- Use context managers or pathlib for one-shot file access (SIM115)
- Collapse nested ifs, drop mutable argument defaults, fix naive
  datetime usage, and clean up the remaining one-off findings
  (SIM102, B006, DTZ, and others)

Notable behavior fixes found along the way:

- tools/waf/pebble_test.py: test_libs mutable default accumulated
  "pthread"/"m" across clar() calls
- tools/power_parser.py: duplicate "Mag" dict key shadowed
  MagPowerSystem
- tools/analyze_mcu_flash_usage_treemap.py, tools/roundrect.py: text
  written to files opened in binary mode

Validated with a qemu_flint configure/build, a unit-test run, and the
pulse2/pebble-loghash/tools test suites.

+++ tools/generate_pdcs/pebble_commands.py
@@ -308,7 +308,7 @@ def serialize_frame(frame, duration):
 
 def pack_header(size):
     return pack(
-        "<BBhh", DRAW_COMMAND_VERSION, 0, int(round(size[0])), int(round(size[1]))
+        "<BBhh", DRAW_COMMAND_VERSION, 0, round(size[0]), round(size[1])
     )
